% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood.R
\name{CalcLognormalLikelihood}
\alias{CalcLognormalLikelihood}
\title{CalcLognormalLikelihood}
\usage{
CalcLognormalLikelihood(data, mean_range, CV_range, length.out = 100)
}
\arguments{
\item{data}{A vector of sample observations}

\item{mean_range}{A vector with values of the mean}

\item{CV_range}{A vector with CV values to be tested}

\item{length.out}{If the parameter ranges}
}
\details{
The ranges for mean and CV may be of any arbitrary length. If they are a given as a vector of length 2, the function will
assume that this represents the bounding range of a sequence of values. In this instance, a sequence of length equal to the
length.out parameter will be generated.
}
\examples{
my_likelihood <- CalcLognormalLikelihood(somedata, mean_range = c(50e3, 150e3), CV_range = c(0.3, 1.8))

}
