% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Params.R
\name{GammaParams}
\alias{GammaParams}
\title{GammaParams}
\usage{
GammaParams(Mean, CV)
}
\arguments{
\item{Mean}{The mean is a positive numeric value}

\item{CV}{The CV is a positive numeric value}
}
\value{
Returns a named list with elements for alpha and beta
}
\description{
This function will translate mean and CV values into corresponding alpha and beta value for a gamma distribution.
}
\details{
Add some remarks about the parameterization.
}
\examples{
GammaParams(10e3, 1.5)

}
