% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Params.R
\name{LognormalParams}
\alias{LognormalParams}
\title{LognormalParams}
\usage{
LognormalParams(Mean, CV)
}
\arguments{
\item{Mean}{The mean is a positive numeric value}

\item{CV}{The CV is a positive numeric value}
}
\value{
Returns a named list with elements for mu and sigma
}
\description{
This function will translate mean and CV values into corresponding mu and sigma values for a lognormal distribution.
}
\details{
Add some remarks about the parameterization.
}
\examples{
LognormalParams(10e3, 1.5)

}
